+++ NamedPipeWrapper/IO/PipeStreamWrapper.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Runtime.Serialization;
 using System.Text;
 
-namespace NamedPipeWrapper.IO
+namespace ricaun.NamedPipeWrapper.IO
 {
     /// <summary>
     /// Wraps a <see cref="PipeStream"/> object to read and write .NET CLR objects.