+++ NamedPipeWrapper/NamedPipeClient.cs
@@ -1,5 +1,5 @@
-﻿using NamedPipeWrapper.IO;
-using NamedPipeWrapper.Threading;
+﻿using ricaun.NamedPipeWrapper.IO;
+using ricaun.NamedPipeWrapper.Threading;
 using System;
 using System.Collections.Generic;
 using System.IO.Pipes;
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 
-namespace NamedPipeWrapper
+namespace ricaun.NamedPipeWrapper
 {
     /// <summary>
     /// Wraps a <see cref="NamedPipeClientStream"/>.