+++ NamedPipeWrapper/Threading/Worker.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NamedPipeWrapper.Threading
+namespace ricaun.NamedPipeWrapper.Threading
 {
     class Worker
     {