+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] / 2024-09-26
+### Features
+- Support `IJsonService` to use `NewtonsoftJsonService` if available.
+- Release nuget [ricaun.NamedPipeWrapper.Json](https://www.nuget.org/packages/ricaun.NamedPipeWrapper.Json).
+### Updated
+- Update namespace to `ricaun.NamedPipeWrapper`.
+- Update package to `ricaun.NamedPipeWrapper.Json`.
+### Removed
+- Remove `Newtonsoft.Json` reference.
+
 ## [1.7.0] / 2024-01-16
 ### Features
 - Add `net8.0` support.
@@ -40,6 +50,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Test` project
 
 [vNext]: ../../compare/1.5.0...HEAD
+[1.8.0]: ../../compare/1.7.0...1.8.0
 [1.7.0]: ../../compare/1.6.0...1.7.0
 [1.6.0]: ../../compare/1.5.3...1.6.0
 [1.5.3]: ../../compare/1.5.2...1.5.3